% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-contracts.R
\name{analysis_rules}
\alias{analysis_rules}
\title{Define analysis method rules}
\usage{
analysis_rules(...)
}
\arguments{
\item{...}{Two-sided formulas of selector to concrete method specification.}
}
\value{
An \code{analysis_rules} object.
}
\description{
Define analysis method rules
}
