% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-adapters.R
\name{as_flextable}
\alias{as_flextable}
\title{Render a bq table with flextable}
\usage{
as_flextable(x, ...)
}
\arguments{
\item{x}{A \code{bq_table}.}

\item{...}{Additional arguments passed to \code{\link[flextable:flextable]{flextable::flextable()}}.}
}
\value{
A \code{flextable}.
}
\description{
Render a bq table with flextable
}
