% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-adapters.R
\name{as_gt}
\alias{as_gt}
\title{Render a bq table with gt}
\usage{
as_gt(x, ...)
}
\arguments{
\item{x}{A \code{bq_table}.}

\item{...}{Additional arguments passed to \code{\link[gt:gt]{gt::gt()}}.}
}
\value{
A \code{gt_tbl}.
}
\description{
Render a bq table with gt
}
