% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-methods.R
\name{as_tibble.bq_data}
\alias{as_tibble.bq_data}
\title{Drop analytic metadata and return a plain tibble}
\usage{
\method{as_tibble}{bq_data}(x, ...)
}
\arguments{
\item{x}{A \code{bq_data} object.}

\item{...}{Ignored.}
}
\value{
A tibble.
}
\description{
Drop analytic metadata and return a plain tibble
}
