% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqreport-package.R
\docType{package}
\name{bqreport-package}
\alias{bqreport}
\alias{bqreport-package}
\title{bqreport: Reproducible Biomedical Data Analysis}
\description{
Provides metadata-aware planning and execution tools for reproducible biomedical data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bqmaks/bqreport}
  \item Report bugs at \url{https://github.com/bqmaks/bqreport/issues}
}

}
\author{
\strong{Maintainer}: bqmaks \email{qmaxksmu@gmail.com}

Authors:
\itemize{
  \item bqmaks \email{qmaxksmu@gmail.com}
}

}
\keyword{internal}
