% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-comparisons.R
\name{brunner_munzel}
\alias{brunner_munzel}
\title{Construct a Brunner-Munzel two-group method}
\usage{
brunner_munzel(backend)
}
\arguments{
\item{backend}{An explicit backend specification from
\code{\link[=brunnermunzel_backend]{brunnermunzel_backend()}} or \code{\link[=lawstat_backend]{lawstat_backend()}}.}
}
\value{
A \code{two_group_method_spec}.
}
\description{
Construct a Brunner-Munzel two-group method
}
