% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-comparisons.R
\name{brunnermunzel_backend}
\alias{brunnermunzel_backend}
\alias{lawstat_backend}
\title{Construct explicit Brunner-Munzel backends}
\usage{
brunnermunzel_backend(permutation)

lawstat_backend()
}
\arguments{
\item{permutation}{Whether to use the permutation implementation. This is
explicit because it changes inference and computational cost.}
}
\value{
A \code{brunner_munzel_backend_spec}.
}
\description{
Construct explicit Brunner-Munzel backends
}
