% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-analysis.R
\name{common_baseline}
\alias{common_baseline}
\alias{stratified_baseline}
\title{Construct Cox baseline-hazard strategies}
\usage{
common_baseline()

stratified_baseline(by)
}
\arguments{
\item{by}{One or more columns defining separate baseline hazards.}
}
\value{
A \code{cox_baseline_strategy}.
}
\description{
Construct Cox baseline-hazard strategies
}
