% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{confidence_intervals}
\alias{confidence_intervals}
\title{Construct a confidence-interval post-processing step}
\usage{
confidence_intervals(level, method)
}
\arguments{
\item{level}{Confidence level.}

\item{method}{Explicit interval method identifier.}
}
\value{
A \code{model_postprocessing_step}.
}
\description{
Construct a confidence-interval post-processing step
}
