% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-result.R, R/contrasts.R
\name{contrasts.analysis_result}
\alias{contrasts.analysis_result}
\alias{contrasts}
\title{Access registered or computed contrasts}
\usage{
\method{contrasts}{analysis_result}(x)

contrasts(x)
}
\arguments{
\item{x}{A \code{bq_data} or \code{analysis_result}.}
}
\value{
A tidy tibble.
}
\description{
Access registered or computed contrasts
}
