% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlation_output}
\alias{correlation_output}
\title{Construct custom correlation output}
\usage{
correlation_output(
  estimate,
  std_error,
  std_error_scale,
  conf_low,
  conf_high,
  statistic,
  df,
  p_value
)
}
\arguments{
\item{estimate, std_error, conf_low, conf_high, statistic, df, p_value}{Numeric scalars.}

\item{std_error_scale}{Scale on which \code{std_error} is defined.}
}
\value{
A validated \code{correlation_method_output}.
}
\description{
Construct custom correlation output
}
