% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-comparisons.R
\name{cramers_v}
\alias{cramers_v}
\title{Construct Cramer's V effect size}
\usage{
cramers_v(ci_method)
}
\arguments{
\item{ci_method}{Explicit confidence-interval method. Currently
\code{noncentral_chi_squared} is supported.}
}
\value{
An \code{effect_size_spec}.
}
\description{
Construct Cramer's V effect size
}
