% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal-design.R
\name{designs}
\alias{designs}
\title{Access registered study designs}
\usage{
designs(x)
}
\arguments{
\item{x}{A \code{bq_data} object.}
}
\value{
A tidy design registry with current column names.
}
\description{
Access registered study designs
}
