% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine-gray.R
\name{fine_gray_model}
\alias{fine_gray_model}
\title{Construct Fine--Gray competing-risks regression}
\usage{
fine_gray_model(cause, exponentiate = TRUE)
}
\arguments{
\item{cause}{Non-censoring event value defining the cumulative-incidence
estimand.}

\item{exponentiate}{Return subdistribution hazard ratios instead of log
subdistribution hazards.}
}
\value{
A concrete Fine--Gray method specification.
}
\description{
Construct Fine--Gray competing-risks regression
}
