% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-comparisons.R
\name{fisher_exact_test}
\alias{fisher_exact_test}
\title{Construct a Fisher exact omnibus test}
\usage{
fisher_exact_test(simulate_p_value, replicates = NULL)
}
\arguments{
\item{simulate_p_value}{Whether to estimate the p-value by Monte Carlo.}

\item{replicates}{Number of Monte Carlo replicates. Must be explicitly
supplied when \code{simulate_p_value = TRUE}.}
}
\value{
An \code{omnibus_spec}.
}
\description{
Construct a Fisher exact omnibus test
}
