% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-contracts.R
\name{linear_model}
\alias{linear_model}
\alias{logistic_model}
\title{Construct built-in method specifications}
\usage{
linear_model(ci_method = "t", exponentiate = FALSE)

logistic_model(ci_method = "wald", exponentiate = TRUE)
}
\arguments{
\item{ci_method}{Confidence interval method.}

\item{exponentiate}{Whether estimates and compatible contrasts are returned
on the exponentiated ratio scale.}
}
\value{
A concrete \code{method_spec}.
}
\description{
Construct built-in method specifications
}
