% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{no_post_hoc}
\alias{no_post_hoc}
\title{Declare that no post-hoc comparisons are requested}
\usage{
no_post_hoc()
}
\value{
A \code{post_hoc_spec}.
}
\description{
Declare that no post-hoc comparisons are requested
}
