% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-analysis.R
\name{no_subgroup_analysis}
\alias{no_subgroup_analysis}
\alias{joint_interaction}
\alias{separate_subgroup_models}
\title{Construct Cox subgroup-fitting strategies}
\usage{
no_subgroup_analysis()

joint_interaction(modifier)

separate_subgroup_models(by)
}
\arguments{
\item{modifier}{A single effect-modifier column.}

\item{by}{A single column defining independently fitted subsets.}
}
\value{
A \code{cox_subgroup_strategy}.
}
\description{
Construct Cox subgroup-fitting strategies
}
