% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{p_adjustment}
\alias{p_adjustment}
\alias{built_in_multiplicity}
\alias{no_adjustment}
\title{Construct an explicit multiplicity specification}
\usage{
p_adjustment(method)

built_in_multiplicity()

no_adjustment()
}
\arguments{
\item{method}{A method accepted by \code{\link[stats:p.adjust]{stats::p.adjust()}}.}
}
\value{
A \code{multiplicity_spec}.
}
\description{
Construct an explicit multiplicity specification
}
