% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{pairwise_post_hoc}
\alias{pairwise_post_hoc}
\title{Construct a pairwise post-hoc step}
\usage{
pairwise_post_hoc(method, comparisons, multiplicity)
}
\arguments{
\item{method}{A post-hoc method created by \code{\link[=tukey_test]{tukey_test()}},
\code{\link[=games_howell_test]{games_howell_test()}}, or \code{\link[=dunn_test]{dunn_test()}}.}

\item{comparisons}{An explicit \code{contrast_spec} identifying target pairs.}

\item{multiplicity}{A multiplicity specification created by
\code{\link[=p_adjustment]{p_adjustment()}} or \code{\link[=built_in_multiplicity]{built_in_multiplicity()}}.}
}
\value{
A \code{post_hoc_spec} and \code{model_postprocessing_step}.
}
\description{
Construct a pairwise post-hoc step
}
