% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized-cox.R
\name{penalized_cox_model}
\alias{penalized_cox_model}
\title{Construct cross-validated penalized Cox regression}
\usage{
penalized_cox_model(
  alpha = 1,
  lambda = c("lambda.min", "lambda.1se"),
  nfolds = 10L,
  seed
)
}
\arguments{
\item{alpha}{Elastic-net mixing parameter between zero and one.}

\item{lambda}{Selected cross-validation solution: \code{lambda.min} or
\code{lambda.1se}.}

\item{nfolds}{Number of cross-validation folds.}

\item{seed}{Explicit reproducibility seed.}
}
\value{
A concrete penalized Cox method specification.
}
\description{
Construct cross-validated penalized Cox regression
}
