% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{per}
\alias{per}
\alias{log2_transform}
\alias{log10_transform}
\title{Scalar transformation specifications}
\usage{
per(increment)

log2_transform()

log10_transform()
}
\arguments{
\item{increment}{Positive increment represented by one coefficient unit.}
}
\value{
A \code{transformation_spec}.
}
\description{
Scalar transformation specifications
}
