% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine-gray.R
\name{plan_fine_gray}
\alias{plan_fine_gray}
\title{Compile a Fine--Gray regression plan}
\usage{
plan_fine_gray(
  .data,
  outcomes = tidyselect::everything(),
  predictors = all_predictors(),
  covariates = tidyselect::any_of(character()),
  confidence_level = 0.95,
  method
)
}
\arguments{
\item{.data}{A \code{bq_data} object.}

\item{outcomes}{Composite competing-risk outcomes selected with tidyselect.}

\item{predictors}{Predictor columns selected with tidyselect.}

\item{covariates}{Optional adjustment covariates selected with tidyselect.}

\item{confidence_level}{Confidence level.}

\item{method}{A \code{\link[=fine_gray_model]{fine_gray_model()}} specification containing the target
event cause.}
}
\value{
An \code{analysis_plan} tibble.
}
\description{
Compile a Fine--Gray regression plan
}
