% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized-cox.R
\name{plan_penalized_cox}
\alias{plan_penalized_cox}
\title{Compile penalized Cox regression tasks}
\usage{
plan_penalized_cox(
  .data,
  outcomes = tidyselect::everything(),
  predictors = all_predictors(),
  covariates = tidyselect::any_of(character()),
  confidence_level = 0.95,
  method
)
}
\arguments{
\item{.data}{A \code{bq_data} object.}

\item{outcomes}{Composite survival outcomes selected with tidyselect syntax.}

\item{predictors}{Predictor columns selected with tidyselect syntax.}

\item{covariates}{Optional adjustment covariates selected with tidyselect.}

\item{confidence_level}{Confidence level.}

\item{method}{A \code{\link[=penalized_cox_model]{penalized_cox_model()}} specification.}
}
\value{
An \code{analysis_plan} tibble.
}
\description{
Compile penalized Cox regression tasks
}
