% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-plots.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot a correlation heatmap}
\usage{
plot_correlation(x)
}
\arguments{
\item{x}{An \code{analysis_result}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a correlation heatmap
}
