% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-plots.R
\name{plot_forest}
\alias{plot_forest}
\title{Plot model estimates and confidence intervals}
\usage{
plot_forest(x, component = c("estimates", "contrasts"))
}
\arguments{
\item{x}{An \code{analysis_result}.}

\item{component}{Either model \code{estimates} or \code{contrasts}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot model estimates and confidence intervals
}
