% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-plots.R
\name{plot_longitudinal}
\alias{plot_longitudinal}
\title{Plot longitudinal change contrasts}
\usage{
plot_longitudinal(
  x,
  estimand = c("change_from_baseline", "difference_in_changes"),
  locale = "en"
)
}
\arguments{
\item{x}{An \code{analysis_result}.}

\item{estimand}{Longitudinal contrast estimand.}

\item{locale}{Output locale, \code{en} or \code{ru}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot longitudinal change contrasts
}
