% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting-plots.R
\name{plot_survival}
\alias{plot_survival}
\title{Plot survival or cumulative-incidence curves}
\usage{
plot_survival(
  x,
  estimand = c("survival", "cumulative_risk", "cumulative_incidence"),
  locale = "en"
)
}
\arguments{
\item{x}{An \code{analysis_result}.}

\item{estimand}{Curve estimand to display.}

\item{locale}{Output locale, \code{en} or \code{ru}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot survival or cumulative-incidence curves
}
