% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-contracts.R
\name{poisson_model}
\alias{poisson_model}
\title{Construct a Poisson count-regression specification}
\usage{
poisson_model(ci_method = "wald", exponentiate = TRUE)
}
\arguments{
\item{ci_method}{Confidence interval method.}

\item{exponentiate}{Whether estimates and compatible contrasts are returned
on the exponentiated ratio scale.}
}
\value{
A concrete \code{method_spec} returning rate ratios by default.
}
\description{
Construct a Poisson count-regression specification
}
