% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-terms.R
\name{polynomial_term}
\alias{polynomial_term}
\title{Construct a polynomial covariate term}
\usage{
polynomial_term(degree, raw = FALSE)
}
\arguments{
\item{degree}{Positive polynomial degree.}

\item{raw}{Whether to use raw rather than orthogonal polynomials.}
}
\value{
A \code{model_term_spec}.
}
\description{
Construct a polynomial covariate term
}
