% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-results.R
\name{print.bq_preflight}
\alias{print.bq_preflight}
\title{Print a preflight result}
\usage{
\method{print}{bq_preflight}(x, ...)
}
\arguments{
\item{x}{A \code{bq_preflight} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a preflight result
}
