% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-results.R
\name{print.bq_result}
\alias{print.bq_result}
\title{Print an analysis result}
\usage{
\method{print}{bq_result}(x, ...)
}
\arguments{
\item{x}{A \code{bq_result} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an analysis result
}
