% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-comparison.R
\name{print.bq_result_comparison}
\alias{print.bq_result_comparison}
\title{Print a comparison result}
\usage{
\method{print}{bq_result_comparison}(x, ...)
}
\arguments{
\item{x}{A \code{bq_result_comparison} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a comparison result
}
