% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-data.R
\name{remove_role}
\alias{remove_role}
\title{Remove an analytical role from variables}
\usage{
remove_role(.data, .cols, role)
}
\arguments{
\item{.data}{A \code{bq_data} object.}

\item{.cols}{Columns selected using tidyselect syntax.}

\item{role}{A single supported role.}
}
\value{
\code{.data} with an updated variable registry.
}
\description{
If the last substantive role is removed, the variable is assigned the
default \code{auxiliary} role.
}
