% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{set_cluster}
\alias{set_cluster}
\title{Configure an observation cluster}
\usage{
set_cluster(.data, .cols, type = "matched_set")
}
\arguments{
\item{.data}{A \code{bq_data} object.}

\item{.cols}{Exactly one cluster identifier selected with tidyselect.}

\item{type}{Cluster semantics. Currently matched sets are supported.}
}
\value{
Updated \code{bq_data}.
}
\description{
Configure an observation cluster
}
