% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{set_coding}
\alias{set_coding}
\title{Configure model coding}
\usage{
set_coding(.data, .cols, coding = "treatment", reference)
}
\arguments{
\item{.data}{A \code{bq_data} object.}

\item{.cols}{Categorical predictors selected with tidyselect.}

\item{coding}{Currently only treatment coding is supported.}

\item{reference}{Reference value.}
}
\value{
Updated \code{bq_data}.
}
\description{
Configure model coding
}
