% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{set_comparisons}
\alias{set_comparisons}
\title{Register target comparisons}
\usage{
set_comparisons(.data, .cols, comparisons, adjust = "none")
}
\arguments{
\item{.data}{A \code{bq_data} object.}

\item{.cols}{Predictor columns selected with tidyselect.}

\item{comparisons}{A \code{contrast_spec}.}

\item{adjust}{A method accepted by \code{\link[stats:p.adjust]{stats::p.adjust()}}.}
}
\value{
Updated \code{bq_data}.
}
\description{
Register target comparisons
}
