% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{set_transformation}
\alias{set_transformation}
\title{Assign scalar transformations}
\usage{
set_transformation(.data, .cols, transformation)
}
\arguments{
\item{.data}{A \code{bq_data} object.}

\item{.cols}{Numeric predictors or covariates selected with tidyselect.}

\item{transformation}{A \code{transformation_spec}.}
}
\value{
Updated \code{bq_data}.
}
\description{
Assign scalar transformations
}
