% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{set_weight}
\alias{set_weight}
\title{Configure a regression weight}
\usage{
set_weight(.data, .cols, type = c("ipw", "frequency", "precision"))
}
\arguments{
\item{.data}{A \code{bq_data} object.}

\item{.cols}{Exactly one weight column selected with tidyselect.}

\item{type}{Weight semantics: frequency, precision, or IPW.}
}
\value{
Updated \code{bq_data}.
}
\description{
Configure a regression weight
}
