% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-descriptive.R
\name{table_body}
\alias{table_body}
\alias{table_header}
\title{Access a backend-independent table component}
\usage{
table_body(x)

table_header(x)
}
\arguments{
\item{x}{A \code{bq_table} object.}
}
\value{
A tibble.
}
\description{
Access a backend-independent table component
}
