% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{tukey_test}
\alias{tukey_test}
\alias{games_howell_test}
\alias{dunn_test}
\title{Construct built-in post-hoc method specifications}
\usage{
tukey_test()

games_howell_test()

dunn_test()
}
\value{
A \code{post_hoc_method_spec}.
}
\description{
Construct built-in post-hoc method specifications
}
