% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-settings.R
\name{variable_colors}
\alias{variable_colors}
\title{Resolve colors registered for a categorical variable}
\usage{
variable_colors(.data, .col)
}
\arguments{
\item{.data}{A \code{bq_data} object.}

\item{.col}{Exactly one column selected with tidyselect.}
}
\value{
A named character vector or \code{NULL}.
}
\description{
Resolve colors registered for a categorical variable
}
