% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-data.R
\name{variables}
\alias{variables}
\title{Access the variable registry}
\usage{
variables(x)
}
\arguments{
\item{x}{A \code{bq_data} object.}
}
\value{
A tibble containing variable metadata.
}
\description{
Access the variable registry
}
