% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{weighted_pearson_correlation}
\alias{weighted_pearson_correlation}
\alias{repeated_measures_correlation}
\title{Construct weighted and repeated-measures correlation methods}
\usage{
weighted_pearson_correlation()

repeated_measures_correlation()
}
\value{
A concrete \code{correlation_method_spec}.
}
\description{
Construct weighted and repeated-measures correlation methods
}
