% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{where_distribution}
\alias{where_distribution}
\alias{where_gaussian}
\alias{where_skewed}
\title{Select variables by distribution profile}
\usage{
where_distribution(distribution)

where_gaussian()

where_skewed()
}
\arguments{
\item{distribution}{One or more distribution profile identifiers.}
}
\value{
A tidyselect selection.
}
\description{
Select variables by distribution profile
}
