% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{where_inferred}
\alias{where_inferred}
\title{Select variables whose type was inferred}
\usage{
where_inferred()
}
\value{
A tidyselect selection.
}
\description{
Select variables whose type was inferred
}
