% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{where_role}
\alias{where_role}
\alias{all_outcomes}
\alias{all_predictors}
\alias{all_groups}
\title{Select variables by analytical role}
\usage{
where_role(role)

all_outcomes()

all_predictors()

all_groups()
}
\arguments{
\item{role}{A supported analytical role.}
}
\value{
A tidyselect selection.
}
\description{
These helpers select variables solely from registry metadata. They do not
filter variables by validation status and can be composed with ordinary
tidyselect expressions.
}
