% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{where_status}
\alias{where_status}
\title{Select variables by registry status}
\usage{
where_status(status)
}
\arguments{
\item{status}{One or more registry status values.}
}
\value{
A tidyselect selection.
}
\description{
Select variables by registry status
}
