% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{where_type}
\alias{where_type}
\alias{where_binary}
\alias{where_continuous}
\alias{where_ordinal}
\alias{where_nominal}
\alias{where_count}
\title{Select variables by analytical type}
\usage{
where_type(type)

where_binary()

where_continuous()

where_ordinal()

where_nominal()

where_count()
}
\arguments{
\item{type}{One or more supported analytical variable types.}
}
\value{
A tidyselect selection.
}
\description{
Select variables by analytical type
}
