% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq-data.R
\name{as_bq_data}
\alias{as_bq_data}
\title{Convert data to a bq data object}
\usage{
as_bq_data(x, metadata = NULL)
}
\arguments{
\item{x}{A data frame.}

\item{metadata}{An optional metadata data frame. See \code{\link[=apply_dictionary]{apply_dictionary()}}.}
}
\value{
A \code{bq_data} tibble.
}
\description{
\code{as_bq_data()} creates a tibble subclass and initializes the package's
metadata registries. Existing \code{bq_data} objects are returned without
regenerating variable identifiers.
}
