% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{biweight_correlation}
\alias{biweight_correlation}
\alias{polychoric_correlation}
\alias{tetrachoric_correlation}
\title{Construct robust and latent-correlation methods}
\usage{
biweight_correlation()

polychoric_correlation()

tetrachoric_correlation()
}
\value{
A concrete \code{correlation_method_spec}.
}
\description{
Construct robust and latent-correlation methods
}
