% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R, R/result-accessors.R
\name{contrasts}
\alias{contrasts}
\alias{contrasts.analysis_result}
\title{Access registered or computed contrasts}
\usage{
contrasts(x)

\method{contrasts}{analysis_result}(x)
}
\arguments{
\item{x}{A \code{bq_data} or \code{analysis_result}.}
}
\value{
A tidy tibble.
}
\description{
Access registered or computed contrasts
}
