% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-methods.R
\name{correlation_comparison_output}
\alias{correlation_comparison_output}
\title{Construct correlation comparison output}
\usage{
correlation_comparison_output(statistic, df, p_value, contrasts)
}
\arguments{
\item{statistic, df, p_value}{Omnibus test values.}

\item{contrasts}{A data frame containing pairwise comparison estimates.}
}
\value{
A \code{correlation_comparison_output}.
}
\description{
Construct correlation comparison output
}
