% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal-analysis.R
\name{lmm_model}
\alias{lmm_model}
\alias{gee_model}
\alias{glmm_model}
\alias{binary_gee_model}
\title{Construct longitudinal model specifications}
\usage{
lmm_model(reml = FALSE)

gee_model(correlation = c("exchangeable", "independence", "ar1"))

glmm_model()

binary_gee_model(correlation = c("exchangeable", "independence", "ar1"))
}
\arguments{
\item{reml}{Whether LMM estimation uses REML. Defaults to FALSE so nested
fixed-effect tests are comparable.}

\item{correlation}{Working correlation structure for GEE.}
}
\value{
A concrete longitudinal method specification.
}
\description{
Construct longitudinal model specifications
}
