% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{pearson_correlation}
\alias{pearson_correlation}
\alias{spearman_correlation}
\alias{kendall_correlation}
\title{Construct correlation methods}
\usage{
pearson_correlation()

spearman_correlation()

kendall_correlation()
}
\value{
A concrete \code{correlation_method_spec}.
}
\description{
Construct correlation methods
}
