% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-comparisons.R
\name{t_test}
\alias{t_test}
\title{Construct a Student or Welch two-sample t method}
\usage{
t_test(var_equal)
}
\arguments{
\item{var_equal}{Whether to use a pooled-variance Student t procedure.
This decision must be supplied explicitly.}
}
\value{
A \code{two_group_method_spec}.
}
\description{
Construct a Student or Welch two-sample t method
}
